backend/requirements_flexible.txt
annotated-types
anyio
beautifulsoup4
blis
cachetools
catalogue
certifi
charset-normalizer
click
cloudpathlib
confection
cymem
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi
filelock
fsspec
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
h11
hf-xet
httplib2
huggingface-hub
idna
Jinja2
langcodes
language_data
marisa-trie
markdown-it-py
MarkupSafe
mdurl
murmurhash
numpy
oauthlib
packaging
preshed
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pyparsing
python-dotenv
PyYAML
regex
requests
requests-oauthlib
rich
rsa
safetensors
scipy
setuptools
shellingham
smart-open
sniffio
soupsieve
spacy
spacy-legacy
spacy-loggers
srsly
starlette
thinc
tokenizers
tqdm
typer
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
wasabi
weasel
wrapt
